fix(paths): normalize separators after home-token expansion; consolidate impl
`$HOME/x` and `${HOME}/x` went through `string.Replace`, which preserves the
literal `/` from the source. `~/x` went through `Path.Combine`, which emits
`Path.DirectorySeparatorChar`. On Windows that produced visually inconsistent
output across the four token forms (`C:\Users\you/x` vs `C:\Users\you\x`) and
broke `NetclawPathsTests` on windows-latest while still passing on Linux.

Collapse `Netclaw.Configuration.NetclawPaths.ExpandHome` and
`Netclaw.Security.PathUtility.ExpandHome` into a single canonical
implementation in `Netclaw.Configuration.PathExpansion`. Both call sites
delegate; `PathUtility.ExpandHome` preserves its non-null contract via a
`?? path` fallback. The canonical impl rewrites `/` to
`Path.DirectorySeparatorChar` only when an expansion actually occurred, so
inputs with no tokens (e.g. `/absolute/path`) still round-trip verbatim.
